Project Description
Snowman Attack displays one of two snowmen for 1 second (as it stands right now). It does this by randomly picking a number 1 or 2.  If a 1 is chosen the good snowman will be displayed at a random location.  If a 2 is choses the bad snowman will be displayed at a random location.  The goal for the user is to click on the bad snowmen only.  If the user clicks on a good snowman the player loses 10 points.  If the user clicks on a bad snowman the player gains 10 points.  If the user just clicks on the background, the player loses 5 points.  Each time the user clicks, the score is updated and a message indicating whether he
or she hit a good snowman, hit a bad snowman, or missed completely.