Edge of Extinction Download Link
A large-scaled C# project built with a team of 3 fellow peers in the Unity Engine for game development. This was a capstone project where we were assigned a client, and needed to go through the entire software development life cycle process. Attached in the zip folder of the project when downloaded are various folders containing our final build of the game - that we showcased to the class in a formal presentation at the end of the semester. As well as important documentation written by myself such as our test plan and technical overview for the client. 

To play the game, simply download the zip folder, unzip it, and go into the 'gamebuild' folder. There you should find an executable file called 'Edge of Extinction'. Double click the exe and begin playing!